wip commit
wip commit 2

cis test for reductino completes successfully

make focusspectra backwards compatible with norm

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

add missing import

fixed tests

add back peak caching

fix aplynormalizationrecipe test

fixed some tests? pytest complains abouot importing an algo locally

just mock out the data service wholesale

will this patch a mock?

fix souschef test

fix grocery service unit tests

fix test generatefocussedvanadium

fix generic recipe, fix normlization service tests

up test coverage

update docstring

address review comments

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

adjust tests

couple more lines to fix

update cis script

update cis test script to point to right norm verion name

add helpful exception for when prenormalizedByCurrent data incorrectly submitted

+++ tests/unit/backend/service/test_CalibrationService.py
@@ -255,6 +255,8 @@ class TestCalibrationServiceMethods(unittest.TestCase):
         self.instance.dataFactoryService.getCifFilePath = MagicMock(return_value="good/cif/path")
         self.instance._collectMetrics = MagicMock(return_value=fakeMetrics)
 
+        FarmFreshIngredients.return_value.get.return_value = True
+
         # Call the method to test
         request = CalibrationAssessmentRequest(
             workspaces={
@@ -314,6 +316,10 @@ class TestCalibrationServiceMethods(unittest.TestCase):
 
             # Under a mocked calibration data path, create fake "persistent" workspace files
             self.instance.dataFactoryService.getCalibrationDataPath = MagicMock(return_value=tmpDir)
+            self.instance.groceryService.dataService.readCalibrationRecord = MagicMock()
+            self.instance.groceryService.fetchCalibrationWorkspaces = MagicMock()
+            self.instance.groceryService._createDiffcalTableWorkspaceName = MagicMock()
+            self.instance.groceryService.fetchGroceryDict = MagicMock()
             self.create_fake_diffcal_files(Path(tmpDir), calibRecord.workspaces, calibRecord.version)
 
             mockRequest = MagicMock(runId=calibRecord.runNumber, version=calibRecord.version, checkExistent=False)
@@ -326,7 +332,8 @@ class TestCalibrationServiceMethods(unittest.TestCase):
             # Delete any existing _data_ workspaces:
             for wss in calibRecord.workspaces.values():
                 for ws in wss:
-                    DeleteWorkspace(Workspace=ws)
+                    if mtd.doesExist(ws):
+                        DeleteWorkspace(Workspace=ws)
 
             mockRequest = MagicMock(runId=calibRecord.runNumber, version=calibRecord.version, checkExistent=True)
             with pytest.raises(RuntimeError) as excinfo:  # noqa: PT011
@@ -344,6 +351,7 @@ class TestCalibrationServiceMethods(unittest.TestCase):
             self.create_fake_diffcal_files(Path(tmpDir), calibRecord.workspaces, calibRecord.version)
 
             mockRequest = MagicMock(runId=calibRecord.runNumber, version=calibRecord.version, checkExistent=False)
+            self.instance.groceryService.dataService.readCalibrationRecord = MagicMock(return_value=calibRecord)
             self.instance.groceryService._getCalibrationDataPath = MagicMock(return_value=tmpDir)
             self.instance.groceryService._fetchInstrumentDonor = MagicMock(return_value=self.sampleWS)
 
@@ -374,6 +382,7 @@ class TestCalibrationServiceMethods(unittest.TestCase):
             mockRequest = MagicMock(
                 runId=calibRecord.runNumber, useLiteMode=True, version=calibRecord.version, checkExistent=False
             )
+            self.instance.groceryService.dataService.readCalibrationRecord = MagicMock(return_value=calibRecord)
             self.instance.groceryService._getCalibrationDataPath = MagicMock(return_value=tmpDir)
             self.instance.groceryService._fetchInstrumentDonor = MagicMock(return_value=self.sampleWS)
             self.instance.loadQualityAssessment(mockRequest)
